packages/py-core/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio[asgi]==5.10.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
rq==1.15.1
cryptography==41.0.7
playwright==1.40.0
pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2

